feat: Add Technologies component
Add a new component called Technologies to display a list of technologies. The Technologies component renders a list of technology images with links to Google search results for each technology. The component is imported and commented out in the Page component, but will be uncommented in a future commit. The Technologies component will receive an array of technology objects as props, which include the name of the technology and an image. The Technologies component is styled using the technologies.css file.